Approval DELETE implemented
also cleaned up bash script STOP output if there
is no container running

+++ reservation-app/bin/dockerAppStop.sh
@@ -13,11 +13,13 @@ echo "Stopping ${service_name} Docker image for ${TEAM} from ${project_dir}"
 
 container_id=`docker ps -aqf name="${TEAM}-${service_name}"`
 
+if [ "${container_id}" != "" ]; then
+    echo "Stopping Docker ID:${container_id}"
+    docker stop ${container_id}
 
-echo "Stopping Docker ID:${container_id}"
-docker stop ${container_id}
+    echo "Removing Docker ID:${container_id}"
+    docker rm ${container_id}
 
-echo "Removing Docker ID:${container_id}"
-docker rm ${container_id}
+    docker ps
+fi
 
-docker ps
