+++ reservation-service/bin/dockerServiceStop.sh
@@ -13,11 +13,15 @@ echo "Stopping ${service_name} Docker image for ${TEAM} from ${project_dir}"
 
 container_id=`docker ps -aqf name="${TEAM}-${service_name}"`
 
+container_id=`docker ps -aqf name="${TEAM}-${service_name}"`
 
-echo "Stopping Docker ID:${container_id}"
-docker stop ${container_id}
+if [ "${container_id}" != "" ]; then
+    echo "Stopping Docker ID:${container_id}"
+    docker stop ${container_id}
 
-echo "Removing Docker ID:${container_id}"
-docker rm ${container_id}
+    echo "Removing Docker ID:${container_id}"
+    docker rm ${container_id}
 
-docker ps
+    docker ps
+fi
+
